chore(client): reorganize pages and components
